+++ API/EnumerableMatcherInterface.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Kaliop\eZMigrationBundle\API;
+
+interface EnumerableMatcherInterface
+{
+    /**
+     * Returns the list of the valid keys to be used in conditions arrays
+     * @return string[]
+     */
+    public function listAllowedConditions();
+}