+++ Core/Executor/RoleManager.php
@@ -7,6 +7,7 @@ use eZ\Publish\API\Repository\RoleService;
 use eZ\Publish\API\Repository\UserService;
 use Kaliop\eZMigrationBundle\API\Collection\RoleCollection;
 use Kaliop\eZMigrationBundle\API\MigrationGeneratorInterface;
+use Kaliop\eZMigrationBundle\API\EnumerableMatcherInterface;
 use Kaliop\eZMigrationBundle\Core\Helper\LimitationConverter;
 use Kaliop\eZMigrationBundle\Core\Matcher\RoleMatcher;
 use eZ\Publish\API\Repository\Values\User\Limitation;
@@ -14,7 +15,7 @@ use eZ\Publish\API\Repository\Values\User\Limitation;
 /**
  * Handles role migrations.
  */
-class RoleManager extends RepositoryExecutor implements MigrationGeneratorInterface
+class RoleManager extends RepositoryExecutor implements MigrationGeneratorInterface, EnumerableMatcherInterface
 {
     protected $supportedStepTypes = array('role');
     protected $supportedActions = array('create', 'load', 'update', 'delete');
@@ -273,6 +274,14 @@ class RoleManager extends RepositoryExecutor implements MigrationGeneratorInterf
 
         $this->loginUser($previousUserId);
         return $data;
+    }
+
+    /**
+     * @return string[]
+     */
+    public function listAllowedConditions()
+    {
+        return $this->roleMatcher->listAllowedConditions();
     }
 
     /**